Rename; restructure; begin adding additional attributes
